feat: user profile: wires up single grid items so that clicking in brings you to that recipe
